server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
simple-websocket==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
psycopg[binary]==3.2.3
python-dotenv==1.0.0